+++ packages/ui/src/components/index.ts
@@ -33,3 +33,4 @@ export * from "./TextInput";
 export * from "./Toggle";
 export * from "./ToggleGroup";
 export * from "./Tooltip";
+export * from "./UIProvider";